+++ packages/map-template/src/components/MapTemplate/MapTemplate.jsx
@@ -234,8 +234,8 @@ function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, p
 
             const miSdkApiTag = document.createElement('script');
             miSdkApiTag.setAttribute('type', 'text/javascript');
-            miSdkApiTag.setAttribute('src', 'https://app.mapsindoors.com/mapsindoors/js/sdk/4.45.0/mapsindoors-4.45.0.js.gz');
-            miSdkApiTag.setAttribute('integrity', 'sha384-g6i/cAWAyDc9wRRgY49L0cz2lPKh8emTU9+fGSIRfxDuQaX3tXA7i3Fgd1kfJ5eJ');
+            miSdkApiTag.setAttribute('src', 'https://app.mapsindoors.com/mapsindoors/js/sdk/4.46.0/mapsindoors-4.46.0.js.gz');
+            miSdkApiTag.setAttribute('integrity', 'sha384-Ygw6sO8Q71z/FxztdpmuOo+hYFf5ozKSIrrm+MDsIaRapLjN7STzmbVNEfHegsqt');
             miSdkApiTag.setAttribute('crossorigin', 'anonymous');
             document.body.appendChild(miSdkApiTag);
             miSdkApiTag.onload = () => {