+++ packages/map-template/CHANGELOG.md
@@ -5,6 +5,48 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.89.1] - 2025-12-11
+
+### Fixed
+
+- Update packages.
+
+## [1.89.0] - 2025-12-11
+
+### Added
+
+- Possibility to remove Wayfinding button via App Config.
+
+## [1.88.1] - 2025-12-11
+
+### Fixed
+
+- Fixed search results sorting to be numeric instead of lexicographic
+
+## [1.88.0] - 2025-12-11
+
+### Added
+
+- Do not show floor in location info when venue has only one floor.
+
+## [1.87.15] - 2025-12-09
+
+### Added
+
+- Added translations for map controls 'Zoom in' and 'Zoom out'
+- Added translations for 'Expand opening hours' and 'Collapse opening hours'
+- Added accessibility translations for text size button
+- Added translation for 'Include bus transportation'
+- Added 'Show legend' translation in i18n files
+
+### Fixed
+
+- Enhanced accessibility by adding aria-labels to buttons in Directions, LocationDetails, OpeningHours, Search, and ShuttleBus components
+
+### Changed
+
+- Removed searchDisabled state and updated disabled logic in Search component for improved accessibility
+
 ## [1.87.14] - 2025-12-04
 
 ### Fixed